requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
psycopg2-binary>=2.9.0
gspread>=5.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0 